\name{mmSeqs}
\alias{mmSeqs}
\docType{data}
\title{ Mouse Mature microRNA Sequences}
\description{
 A set of mouse microRNA sequences.
}
\usage{data(mmSeqs)}
\format{
 A character vector.
}
\details{
 Each sequence represents a different mature mouse microRNA.
}
\source{
\url{http://microrna.sanger.ac.uk/sequences/index.shtml}
}
\references{
miRBase: microRNA sequences, targets and gene nomenclature.
Griffiths-Jones S, Grocock RJ, van Dongen S, Bateman A, Enright AJ.
NAR, 2006, 34, Database Issue, D140-D144

The microRNA Registry.
Griffiths-Jones S.
NAR, 2004, 32, Database Issue, D109-D111

}
\examples{
data(mmSeqs)
}
\keyword{datasets}
